backend/requirements.txt
alembic==1.14.1; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
flask-mail==0.10.0; python_version >= '3.8'
flask-marshmallow==1.3.0; python_version >= '3.9'
flask-migrate==4.1.0; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
imagekitio==4.1.0; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
mako==1.3.9; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
marshmallow-sqlalchemy==1.4.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pendulum==3.0.0; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==0.10.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
schedule==1.2.2; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.38; python_version >= '3.7'
time-machine==2.16.0; implementation_name != 'pypy'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
urllib3==1.26.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==3.1.3; python_version >= '3.9'